Extract run banner marker writer
